Clean templates and update vote view function
All of the html templates now reference the base
html for the head, styling, and navbar code. I
also consolidated all of the styling rules into a
single file for easiers rather than having
multiple templates contain the same styling rules.
The vote function was pointed out to have a minor
oversight on my part, but I fixed it by using a
django built-in function to retrieve a unique poll
and return a 404 error if that poll cannot be
retrieved.